efi: Remove unnecessary cast
clippy suggests removal of unnecessary casting integer literal to `u64`.

+++ src/efi/block.rs
@@ -190,7 +190,7 @@ impl<'a> BlockWrapper<'a> {
             AllocateType::AllocateAnyPages,
             MemoryType::LoaderData,
             ((size + super::PAGE_SIZE as usize - 1) / super::PAGE_SIZE as usize) as u64,
-            0 as u64,
+            0_u64,
         );
 
         let bw = new_address as *mut BlockWrapper;
